+++ exporters/metric/cortex/cortex_test.go
@@ -35,7 +35,7 @@ import (
 	export "go.opentelemetry.io/otel/sdk/export/metric"
 	"go.opentelemetry.io/otel/sdk/export/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/resource"
-	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.7.0"
 )
 
 var testResource = resource.NewWithAttributes(semconv.SchemaURL, attribute.String("R", "V"))