+++ exporters/metric/dogstatsd/internal/statsd/conn_test.go
@@ -38,7 +38,7 @@ import (
 	controller "go.opentelemetry.io/otel/sdk/metric/controller/basic"
 	processor "go.opentelemetry.io/otel/sdk/metric/processor/basic"
 	"go.opentelemetry.io/otel/sdk/resource"
-	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.7.0"
 )
 
 func testMeter(t *testing.T, exp export.Exporter) (context.Context, metric.Meter, *controller.Controller) {