+++ instrumentation/github.com/bradfitz/gomemcache/memcache/otelmemcache/internal/attributes.go
@@ -16,7 +16,7 @@ package internal
 
 import (
 	"go.opentelemetry.io/otel/attribute"
-	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.7.0"
 )
 
 type Operation string