+++ instrumentation/github.com/gocql/gocql/otelgocql/internal/cass.go
@@ -19,7 +19,7 @@ import (
 	"net"
 
 	"go.opentelemetry.io/otel/attribute"
-	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.7.0"
 )
 
 const (